delete unused script
