+++ Packages/ParticlePhysics/Tests/Runtime/GridSearch/Scripts/GridSearchTest.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 77a46efb48a1818458389834e632e5a5
-timeCreated: 1499171793
-licenseType: Pro
+guid: 183667f55cefd1a41b33f79d302f63eb
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
