--- Packages/ParticlePhysics/Tests/Runtime/GridSearch/Scripts/Test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 183667f55cefd1a41b33f79d302f63eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
